refactor: move logger to core
