+++ kiauh/components/klipper/klipper_remove.py
@@ -14,9 +14,9 @@ from components.klipper import KLIPPER_DIR, KLIPPER_ENV_DIR
 from components.klipper.klipper import Klipper
 from components.klipper.klipper_dialogs import print_instance_overview
 from core.instance_manager.instance_manager import InstanceManager
+from core.logger import Logger
 from utils.fs_utils import run_remove_routines
 from utils.input_utils import get_selection_input
-from utils.logger import Logger
 from utils.sys_utils import cmd_sysctl_manage
