+++ kiauh/components/klipper_firmware/menus/klipper_flash_menu.py
@@ -36,10 +36,10 @@ from components.klipper_firmware.menus.klipper_flash_help_menu import (
     KlipperMcuConnectionHelpMenu,
 )
 from core.constants import COLOR_CYAN, COLOR_RED, COLOR_YELLOW, RESET_FORMAT
+from core.logger import DialogType, Logger
 from core.menus import FooterType, Option
 from core.menus.base_menu import BaseMenu
 from utils.input_utils import get_number_input
-from utils.logger import DialogType, Logger
 
 
 # noinspection PyUnusedLocal
