+++ kiauh/components/webui_client/client_utils.py
@@ -22,6 +22,7 @@ from components.webui_client.fluidd_data import FluiddData
 from components.webui_client.mainsail_data import MainsailData
 from core.backup_manager.backup_manager import BackupManager
 from core.constants import COLOR_CYAN, COLOR_YELLOW, RESET_FORMAT
+from core.logger import Logger
 from core.settings.kiauh_settings import KiauhSettings
 from utils import NGINX_CONFD, NGINX_SITES_AVAILABLE
 from utils.common import get_install_status
@@ -29,7 +30,6 @@ from utils.git_utils import (
     get_latest_tag,
     get_latest_unstable_tag,
 )
-from utils.logger import Logger
 from utils.types import ComponentStatus
