+++ kiauh/core/menus/main_menu.py
@@ -33,6 +33,7 @@ from core.constants import (
     COLOR_YELLOW,
     RESET_FORMAT,
 )
+from core.logger import Logger
 from core.menus import FooterType
 from core.menus.advanced_menu import AdvancedMenu
 from core.menus.backup_menu import BackupMenu
@@ -42,7 +43,6 @@ from core.menus.remove_menu import RemoveMenu
 from core.menus.settings_menu import SettingsMenu
 from core.menus.update_menu import UpdateMenu
 from extensions.extensions_menu import ExtensionsMenu
-from utils.logger import Logger
 from utils.types import ComponentStatus, StatusMap, StatusText
