+++ kiauh/extensions/extensions_menu.py
@@ -16,11 +16,11 @@ from pathlib import Path
 from typing import Dict, List, Type
 
 from core.constants import COLOR_CYAN, COLOR_YELLOW, RESET_FORMAT
+from core.logger import Logger
 from core.menus import Option
 from core.menus.base_menu import BaseMenu
 from extensions import EXTENSION_ROOT
 from extensions.base_extension import BaseExtension
-from utils.logger import Logger
 
 
 # noinspection PyUnusedLocal
