+++ kiauh/extensions/obico/moonraker_obico.py
@@ -13,6 +13,7 @@ from pathlib import Path
 from subprocess import CalledProcessError, run
 
 from core.instance_manager.base_instance import BaseInstance
+from core.logger import Logger
 from core.submodules.simple_config_parser.src.simple_config_parser.simple_config_parser import (
     SimpleConfigParser,
 )
@@ -26,7 +27,6 @@ from extensions.obico import (
     OBICO_LOG_NAME,
     OBICO_SERVICE_TEMPLATE,
 )
-from utils.logger import Logger
 
 
 # noinspection PyMethodMayBeStatic
