+++ kiauh/main.py
@@ -7,9 +7,9 @@
 #  This file may be distributed under the terms of the GNU GPLv3 license  #
 # ======================================================================= #
 
+from core.logger import Logger
 from core.menus.main_menu import MainMenu
 from core.settings.kiauh_settings import KiauhSettings
-from utils.logger import Logger
 
 
 def main() -> None:
