+++ kiauh/utils/input_utils.py
@@ -12,8 +12,8 @@ import re
 from typing import Dict, List
 
 from core.constants import COLOR_CYAN, RESET_FORMAT
+from core.logger import Logger
 from utils import INVALID_CHOICE
-from utils.logger import Logger
 
 
 def get_confirm(question: str, default_choice=True, allow_go_back=False) -> bool | None:
